Separated tests that are for reallocation of individuals
